% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_model.R
\name{agg_tuning_res}
\alias{agg_tuning_res}
\title{Aggregate tuning results across folds}
\usage{
agg_tuning_res(tune_res)
}
\arguments{
\item{tune_res}{Parameter grid with tuning results included.}
}
\value{
Aggregated version of tune_res, a data frame.
}
\description{
Aggregate tuning results across folds
}
