% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa_data.R
\name{check_covs}
\alias{check_covs}
\title{Check that covariates exist.}
\usage{
check_covs(x, covs)
}
\arguments{
\item{x}{Data frame}

\item{covs}{Covariate names}
}
\value{
No return value, throws an error if check fails.
}
\description{
Check that covariates exist.
}
