% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa_data.R
\name{check_normed}
\alias{check_normed}
\title{Check any co variate columns}
\usage{
check_normed()
}
\value{
A logical flag indicating if format check passed (TRUE)
}
\description{
Check any co variate columns
}
