% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa_data.R
\name{check_y}
\alias{check_y}
\title{Check outcome column}
\usage{
check_y(y_name, x)
}
\arguments{
\item{y_name}{Name of the outcome column}

\item{x}{The input data frame}
}
\value{
A logical flag indicating if format check passed (TRUE)
}
\description{
Check outcome column
}
