% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_model.R
\name{cnn_mp_layer}
\alias{cnn_mp_layer}
\title{Create 1D CNN and Max Pooled Layers for ConTextNet}
\usage{
cnn_mp_layer(k, f, l_cnn, l_corr, input_layer)
}
\arguments{
\item{k}{Kernel size (int)}

\item{f}{Number of filters (int)}

\item{l_cnn}{Kernel regularizer weight (numeric)}

\item{l_corr}{Correlation regularizer weight (numeric)}

\item{input_layer}{Layer corresponding to word embedding input}
}
\value{
R Keras model layer.
}
\description{
Create 1D CNN and Max Pooled Layers for ConTextNet
}
