% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_model.R
\name{combine_tune_res}
\alias{combine_tune_res}
\title{Combine job array tuning results.}
\usage{
combine_tune_res(folder_path, folder_name, grid)
}
\arguments{
\item{folder_path}{Path to parent directory where folder_name should be
stored.}

\item{folder_name}{Name of directory where model files are saved.}

\item{grid}{Parameter tuning grid.}
}
\value{
No return object, but compiled tuning results will be saved if the
job array was successful.
}
\description{
Combine job array tuning results.
}
