% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_model.R
\name{corr_regularizer}
\alias{corr_regularizer}
\title{Activation Correlation Regularizer}
\usage{
corr_regularizer(weight, f)
}
\arguments{
\item{weight}{Weight to apply to this penalty.}

\item{f}{Number of filters for this convolutional layer.}
}
\value{
The penalty value.
}
\description{
Custom regularizer for CNN layers that applies a penalty proportional to the
max correlation between activations of filters within a convolutional layer.
Note: the nested function structure is necessary for passing in additional
parameters.
}
