% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{create_model_dir}
\alias{create_model_dir}
\title{Create new directory for model files}
\usage{
create_model_dir(name, path = "", override = FALSE)
}
\arguments{
\item{name}{Name to give the directory for this model}

\item{path}{Path to the parent directory}

\item{override}{By default, the directory will only be created if you
give permission interactively. Set override = TRUE to create the directory
without interactive checks. Existing directories will not be overwritten -
a new directory name must be given.}
}
\value{
The location of the directory created
}
\description{
This directory will contain the following sub directories...
}
\examples{
\dontrun{create_model_dir("example")}
}
