% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{create_params}
\alias{create_params}
\title{Create a list to store model parameters (meta-params and final model params)}
\usage{
create_params()
}
\description{
Create a list to store model parameters (meta-params and final model params)
}
