% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_doc_acts.R
\name{get_doc_acts}
\alias{get_doc_acts}
\title{Get Document Level Filter Activation Data Frame}
\usage{
get_doc_acts(model, embeds, params, dat = NULL)
}
\arguments{
\item{model}{R Keras trained model}

\item{embeds}{Matrix of word embeddings to get activations for}

\item{params}{Model parameter list}

\item{dat}{Original data set (only necessary if covariates are used in model)}
}
\value{
Data frame with convolutional layer activations for all input word
embeddings.
}
\description{
Get Document Level Filter Activation Data Frame
}
\examples{
\dontrun{
model <- train_model(imdb_embed)
embeds <- imdb_embed$embeds[imdb_embed$dat$fold == "train", , ]
dat <- imdb_embed$dat[imdb_embed$dat$fold == "train", ]
params <- imdb_embed$params
res <- get_doc_acts(model, embeds, params, dat)
}
}
