% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_output_wts.R
\name{get_output_wts}
\alias{get_output_wts}
\title{Get Output Layer Weights}
\usage{
get_output_wts(model, params)
}
\arguments{
\item{model}{Keras model with a final dense layer named "output"}

\item{params}{Model parameters}
}
\value{
Data frame with output layer weights corresponding to each
convolutional layer filter (and potentially covariates)
}
\description{
Get Output Layer Weights
}
\examples{
\dontrun{
model <- train_model(imdb_embed$dat, imdb_embed$embeds, imdb_embed$params)
params <- imdb_embed$params
get_output_wts(model, params)
}
}
