% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_top_phrases.R
\name{get_phrase}
\alias{get_phrase}
\title{Get Text Phrase from Document Tokens}
\usage{
get_phrase(doc_tokens, phrase_id, k, vocab)
}
\arguments{
\item{doc_tokens}{Tokens for the entire document that the phrase is within}

\item{phrase_id}{Position of the starting character of the phrase}

\item{k}{Phrase length}

\item{vocab}{Vocabulary list from the tokenizer.}
}
\value{
String containing the phrase.
}
\description{
Retrieve a phrase given the starting character within \code{doc_tokens}, the
phrase length \code{k}, and the \code{vocab} list from the tokenizer.
}
\examples{
\dontrun{
get_phrase(imdb_embed$tokens[1, ], 6, 4, imdb_embed$vocab)
}
}
