% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_top_phrases.R
\name{get_phrase_acts}
\alias{get_phrase_acts}
\title{Get Phrase Filter Activations (single layer)}
\usage{
get_phrase_acts(model, k, input, n_filts, max_length)
}
\arguments{
\item{model}{R Keras trained model}

\item{k}{Kernel size (indicating which convolutional layer to consider)}

\item{input}{Model input (either embedding matrix or list with embedding
matrix and covariate matrix)}

\item{n_filts}{The number of filters in this convolutional layer}

\item{max_length}{Number of tokens}
}
\value{
A data frame of phrase filter activations for this convolutional
layer.
}
\description{
Get Phrase Filter Activations (single layer)
}
