% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_model.R
\name{get_phrase_metrics}
\alias{get_phrase_metrics}
\title{Tuning Wrapper for get_top_phrases_quick()}
\usage{
get_phrase_metrics(model, params, embeds, tokens, vocab)
}
\arguments{
\item{model}{Trained model}

\item{params}{List of model parameters used in training \code{model}}

\item{embeds}{Word embeddings}

\item{tokens}{Data frame with tokens corresponding to \code{dat$text}}

\item{vocab}{Named list, where values are token values and names are the
corresponding text}
}
\value{
Short summary of top phrases
}
\description{
Tuning Wrapper for get_top_phrases_quick()
}
