% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_model.R
\name{get_row_list}
\alias{get_row_list}
\title{Get Parameter List from Grid Row}
\usage{
get_row_list(grid_row)
}
\arguments{
\item{grid_row}{Grid row of parameter settings to evaluate}
}
\value{
A list
}
\description{
Get Parameter List from Grid Row
}
