% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_top_phrases.R
\name{get_top_phrases}
\alias{get_top_phrases}
\title{Get Highest Activated Filters}
\usage{
get_top_phrases(phrase_acts, tokens, params, vocab, m = 10)
}
\arguments{
\item{phrase_acts}{Data frame with phrase activations from phrase_acts()}

\item{tokens}{Data frame with tokens corresponding to \code{dat$text}}

\item{params}{List of model parameters used in training \code{model}}

\item{vocab}{Named list, where values are token values and names are the
corresponding text}

\item{m}{Number of top phrases to pull per filter.}
}
\value{
Data frame with top phrases and their activations.
}
\description{
Get Highest Activated Filters
}
\examples{
\dontrun{
model <- train_model(imdb_embed)
tokens <- imdb_embed$tokens
embeds <- imdb_embed$embeds[imdb_embed$dat$fold == "train", , ]
dat <- imdb_embed$dat[imdb_embed$dat$fold == "train", ]
params <- imdb_embed$params
vocab <- imdb_embed$vocab
phrase_acts <- get_phrase_acts(model, embeds, params, dat)
get_top_phrases(tokens, embeds, dat, params, vocab)
}
}
