% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_model.R
\name{init_model}
\alias{init_model}
\title{Initialize ConTextNet Model}
\usage{
init_model(params)
}
\arguments{
\item{params}{List of model parameters - from tune_prep() or directly from
data_prep().}
}
\value{
A compiled keras model.
}
\description{
Initialize ConTextNet Model
}
\examples{
\dontrun{
model_params <- list("n_filts" = list(2), "kern_sizes" = list(c(3, 5)),
                     "lr" = list(0.0001), "lambda_cnn" = list(0),
                     "lambda_corr" = list(0), "lambda_out" = list(0),
                     "epochs" = list(20), "batch_size" = list(32),
                     "covars" = list(NULL))
input_list <- prep_data(x = imdb, y_name = "y", text_name = "text",
                        model_params = model_params, task = "class",
                        folder_name = "example")
res <- embed(input_list)
init_model(res$params)
}
}
