% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_model.R
\name{initialize_model}
\alias{initialize_model}
\title{Initialize ConTextNet Model}
\usage{
initialize_model(params)
}
\arguments{
\item{params}{List of model parameters - from tune_prep() or directly from
data_prep().}
}
\description{
Initialize ConTextNet Model
}
