% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_doc_acts.R
\name{plot_doc_acts}
\alias{plot_doc_acts}
\title{Plot Outcome vs. Document-Level Activations}
\usage{
plot_doc_acts(doc_acts)
}
\arguments{
\item{doc_acts}{Data frame with document-level activations, the output of
get_doc_acts().}
}
\value{
A \code{ggplot} scatterplot with a facet per convolutional filter.
}
\description{
Plot Outcome vs. Document-Level Activations
}
\examples{
\dontrun{
imdb_embed$params$epochs <- 100
model <- train_model(imdb_embed)
embeds <- imdb_embed$embeds[imdb_embed$dat$fold == "train", , ]
dat <- imdb_embed$dat[imdb_embed$dat$fold == "train", ]
params <- imdb_embed$params
plot_doc_acts(get_doc_acts(model, embeds, params, dat))
}
}
