% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{prep_params}
\alias{prep_params}
\title{Create a list to store model parameters (meta-params and final model params)}
\usage{
prep_params(p, tune_method)
}
\arguments{
\item{p}{A list, model_params, passed to prep_data().}

\item{tune_method}{How tuning should be performed: locally, via a Cluster
with Slurm, generically via a shell script, or not at all
("local", "slurm", "shell", or "none")}
}
\description{
Create a list to store model parameters (meta-params and final model params)
}
