% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{scale_vars}
\alias{scale_vars}
\title{Scale covariates to have mean-zero and standard deviation of 1 using training
data.}
\usage{
scale_vars(dat, vars, scale_type)
}
\arguments{
\item{dat}{The input data set. Should include a \code{fold} column with \code{train}.
and \code{test} entries.}

\item{vars}{The names of variables to scale.}

\item{scale_type}{Either "normalize" or "min-max".}
}
\value{
The same data frame, but with the \code{covs} columns scaled.
}
\description{
Scale covariates to have mean-zero and standard deviation of 1 using training
data.
}
\examples{
dat <- as.data.frame(matrix(rnorm(12), ncol = 3))
dat$fold <- "train"
dat2 <- scale_vars(dat, c("V1", "V2"), "min-max")
}
