% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenize.R
\name{tokenize}
\alias{tokenize}
\title{Tokenize input text}
\usage{
tokenize(x, tokenizer)
}
\arguments{
\item{x}{A character vector}

\item{tokenizer}{A tokenizer model from the Hugging Face transformer library.}
}
\description{
Tokenize input text
}
