% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_model.R
\name{tune_model}
\alias{tune_model}
\title{Tune ConTextNet Model}
\usage{
tune_model(dat, embeds, meta_params, grid, tokens, vocab)
}
\arguments{
\item{dat}{Original text data set with outcome \code{y} (assumes full, will auto
discard test set)}

\item{embeds}{Text embeddings for \code{dat$text}}

\item{meta_params}{Meta-parameters for model training}

\item{grid}{Grid of parameter settings to evaluate}

\item{tokens}{Tokens for \code{dat$text}}

\item{vocab}{Vocab map from tokenizer}
}
\value{
Data frame, which is grid with model performances filled in.
}
\description{
Tune ConTextNet Model
}
