% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_model.R
\name{corr_regularizer}
\alias{corr_regularizer}
\title{Activation Correlation Regularizer}
\usage{
corr_regularizer(weight, params)
}
\arguments{
\item{weight}{Weight to apply to this penalty.}

\item{params}{List of model parameters.}
}
\value{
The penalty value.
}
\description{
Custom regularizer for CNN layers that applies a penalty proportional to the
max correlation between activations of filters within a convolutional layer.
Note: the nested function structure is necessary for passing in weight and
params.
}
