% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{create_model_dir}
\alias{create_model_dir}
\title{Create new directory for model files}
\usage{
create_model_dir(name)
}
\arguments{
\item{name}{Name to give the directory for this model}
}
\value{
The location of the directory created
}
\description{
This directory will contain the following subdirectories
}
\examples{
\dontrun{create_model_dir("example")}
}
