% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed.R
\name{embed}
\alias{embed}
\title{Get embeddings of texts}
\usage{
embed(input_list)
}
\arguments{
\item{input_list}{This should be an output of prep_data().}
}
\value{
List, which is \code{input_list} now including the text embeddings, with
the embedding dimension recorded in \code{input_list$params$embed_dim}.
}
\description{
Get embeddings of texts
}
\examples{
\dontrun{
res <- embed(input_list)
}
}
