% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed.R
\name{embed}
\alias{embed}
\title{Get embeddings of texts}
\usage{
embed(input_list)
}
\arguments{
\item{input_list}{This should be an output of prep_data().}
}
\value{
List, which is \code{input_list} now including the text embeddings, with
the embedding dimension recorded in \code{input_list$params$embed_dim}.
}
\description{
Get embeddings of texts
}
\examples{
\dontrun{
param_vals <- list("n_filts" = list(2), "kern_sizes" = list(c(3, 5)),
                   "lr" = list(0.0001), "lambda_cnn" = list(0),
                   "lambda_corr" = list(0), "lambda_out" = list(0),
                   "epochs" = list(20), "batch_size" = list(32),
                   "covars" = list(NULL))
input_list <- prep_data(x = imdb, y_name = "y", text_name = "text",
                        param_vals = param_vals, task = "class",
                        folder_name = "example")
res <- embed(input_list)
}
}
