% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_output_wts.R
\name{get_output_wts}
\alias{get_output_wts}
\title{Get Output Layer Weights}
\usage{
get_output_wts(model, kern_sizes, f)
}
\arguments{
\item{model}{Keras model with a final dense layer named "output"}

\item{kern_sizes}{Array of convolutional layer kernel sizes in the model}

\item{f}{Integer, number of filters per convolutional layer}
}
\value{
Data frame with output layer weights corresponding to each
convolutional layer filter (and potentially covariates)
}
\description{
Get Output Layer Weights
}
\examples{
\dontrun{
model <- train_model(imdb_embed$dat, imdb_embed$embeds, imdb_embed$params)
params <- imdb_embed$params
get_output_wts(model, params$kern_sizes, params$n_filts)
}
}
