% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_top_phrases.R
\name{get_top_phrases_quick}
\alias{get_top_phrases_quick}
\title{Get Quick Summary of Highest Activated Filters}
\usage{
get_top_phrases_quick(phrase_acts, tokens, params, vocab)
}
\arguments{
\item{phrase_acts}{Data frame with phrase activations from phrase_acts()}

\item{tokens}{Data frame with tokens corresponding to \code{dat$text}}

\item{params}{List of model parameters used in training \code{model}}

\item{vocab}{Named list, where values are token values and names are the
corresponding text}
}
\value{
A string listing the top 3 phrases most highly associated with the
top 4 convolutional filters with the highest output layer weight.
}
\description{
Get Quick Summary of Highest Activated Filters
}
