% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_model.R
\name{init_model}
\alias{init_model}
\title{Initialize ConTextNet Model}
\usage{
init_model(params)
}
\arguments{
\item{params}{List of model parameters - from tune_prep() or directly from
data_prep().}
}
\value{
A compiled keras model.
}
\description{
Initialize ConTextNet Model
}
\examples{
\dontrun{
init_model(imdb_embed$params)
}
}
