% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_model.R
\name{prep_cluster_tune}
\alias{prep_cluster_tune}
\title{Prepare shell script for parameter tuning cluster job array}
\usage{
prep_cluster_tune(
  grid_len,
  job_iters,
  input_path,
  conda_env,
  cpu_mem,
  runtime,
  cpu_num,
  partition
)
}
\arguments{
\item{grid_len}{Number of rows in the tuning grid}

\item{job_iters}{Number of grid rows to give to each individual job}

\item{input_path}{Path for saved}

\item{conda_env}{Cluster conda environment where packages are installed}

\item{cpu_mem}{Memory per CPU node to request}

\item{runtime}{Duration of each job}

\item{cpu_num}{Number of CPUs to request for each job}

\item{partition}{Name of cluster partition to use}
}
\description{
Prepare shell script for parameter tuning cluster job array
}
