% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_model.R
\name{train_model}
\alias{train_model}
\title{Train a ConTextNet Model}
\usage{
train_model(dat, embeds, params, run_quiet = FALSE)
}
\arguments{
\item{dat}{Original text data set}

\item{embeds}{Embeddings corresponding to dat$text}

\item{params}{Model parameters}

\item{run_quiet}{Logical indicating whether or not to show model training
progress in the console.}
}
\value{
Trained Keras model.
}
\description{
Train a ConTextNet Model
}
\examples{
\dontrun{train_model(imdb_embed)}
}
