% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_model.R
\name{tune_model}
\alias{tune_model}
\title{Title}
\usage{
tune_model(grid, inputs)
}
\arguments{
\item{grid}{Grid of parameter settings from create_grid()}

\item{inputs}{Model training inputs from embed()}
}
\value{
Data frame, which is grid with model performances filled in.
}
\description{
Title
}
